Got rid most of the warnings for Linux
warning were removed by using diagnostic pragmas for GCC where
applicable and appropriate integer literals in the tests. 
The idea was to get rid of all the mess and hopefully see only 
the useful warnings. Did not help with dealing with the SEGFAULT 
on Linux but at least the build looks better. 

Update #100