Destroy `Node`'s handle before its `Link`
Fixes #659.

Nodes must first delete the handle to the underlying hdf5
resource before deleting their link.
Otherwise (i.e., link is deleted first), hdf5 attempts to
close the file (if this was the last node referencing the
file), which might fail (e.g., when using mpi-io, or close
degree `Semi`) because the handle to the node would still
be around.